pyproject.toml
[project]
name = "survey-ip-metadata"
version = "0.1.0"
description = "Collect GeoIP/ASN/WHOIS/hostname metadata for IP addresses"
authors = [{name="Your Name", email="you@example.com"}]
dependencies = [
  "requests>=2.28",
  "ipwhois>=1.1.0",
  "python-dotenv>=1.0",
  "PyYAML>=6.0",
  "tqdm>=4.65"
]
[tool.poetry.scripts]
survey-ip = "survey_ip.cli:main"
